+++ apps/api/AssetManagement.Api/Data/Migrations/20260208152528_AddApplicationInactiveStatus.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace AssetManagement.Api.Data.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddApplicationInactiveStatus : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+
+        }
+    }
+}